Refactor tests for improved clarity and safety
- Updated assertions in test files to use `bool()` for NaN and Inf checks to enhance readability and prevent potential issues.
- Added metadata handling functions for DataLoader objects to ensure type safety and avoid static analysis complaints.
- Modified tests to reflect changes in return values from model building functions, ensuring consistency in expected outputs.
- Cleaned up test cases by removing unnecessary comments and improving the structure for better maintainability.
- Introduced helper functions for safe numeric coercion, enhancing robustness against various input types in utility functions.
- Added debug scripts for inspecting Plotly layout properties, aiding in troubleshooting interactive plot issues.

+++ src/experiments/hyperparameter_sensitivity.py
@@ -50,7 +50,7 @@ def compute_trajectory_smoothness(trajectory: np.ndarray) -> float:
             angles.append(np.clip(cos_angle, -1.0, 1.0))
     
     # Smoothness = high cosine similarity (aligned updates)
-    smoothness = np.mean(angles) if angles else 1.0
+    smoothness = float(np.mean(angles)) if angles else 1.0
     return smoothness
 
 
@@ -76,8 +76,8 @@ def compute_oscillation_index(trajectory: np.ndarray) -> float:
             if cos_angle < 0:  # Direction reversal
                 reversals += 1
     
-    oscillation_index = reversals / max(1, len(updates) - 1)
-    return oscillation_index
+    oscillation_index = float(reversals) / max(1, len(updates) - 1)
+    return float(oscillation_index)
 
 
 def run_optimizer_trajectory(optimizer, func, grad_func, x0, max_iters=1000, tol=1e-6):
